Move /src/extra to /src/elem/
