+++ examples/arduino/ex04_ard_ctrls/ex04_ard_ctrls.ino
@@ -23,9 +23,9 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XSlider.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XSlider.h"
+#include "elem/XGauge.h"
 
 #include <Adafruit_GFX.h>
