+++ examples/arduino/ex07_ard_slider/ex07_ard_slider.ino
@@ -22,7 +22,7 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XSlider.h"
+#include "elem/XSlider.h"
 
 // Defines for resources
