+++ examples/arduino/ex21_ard_input_pin/ex21_ard_input_pin.ino
@@ -22,9 +22,9 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
-#include "extra/XSlider.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
+#include "elem/XSlider.h"
 
 #include <Adafruit_GFX.h>
