+++ examples/arduino/ex23_m5_input_btn/ex23_m5_input_btn.ino
@@ -22,9 +22,9 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
-#include "extra/XSlider.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
+#include "elem/XSlider.h"
 
 // Ensure config settings are correct for the sketch
 #if !defined(DRV_DISP_M5STACK) || !defined(DRV_TOUCH_M5STACK) || !(GSLC_FEATURE_INPUT)
