+++ examples/arduino_min/ex04_ardmin_ctrls/ex04_ardmin_ctrls.ino
@@ -20,9 +20,9 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XSlider.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XSlider.h"
+#include "elem/XGauge.h"
 
 // Defines for resources
