+++ examples/arduino_min/ex06_ardmin_callback/ex06_ardmin_callback.ino
@@ -24,8 +24,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
 
 #include <math.h>
