+++ examples/arduino_min/ex07_ardmin_slider/ex07_ardmin_slider.ino
@@ -20,7 +20,7 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XSlider.h"
+#include "elem/XSlider.h"
 
 
 // Defines for resources
