+++ examples/arduino_min/ex10_ardmin_textbox/ex10_ardmin_textbox.ino
@@ -20,8 +20,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XSlider.h"
-#include "extra/XTextbox.h"
+#include "elem/XSlider.h"
+#include "elem/XTextbox.h"
 
 
 // Defines for resources
