+++ examples/linux/Makefile
@@ -49,7 +49,7 @@ INCLUDE = -I/usr/local/include
 CFLAGS = $(DEBUG) -Wall $(INCLUDE) -Winline -pipe -g
 LDFLAGS = -L/usr/local/lib
 
-GSLC_CORE = ../../src/GUIslice.c ../../src/GUIslice_config.h ../../src/extra/*.c
+GSLC_CORE = ../../src/GUIslice.c ../../src/GUIslice_config.h ../../src/elem/*.c
 
 
 # Define default compiler flags which may be overridden by command line
