+++ examples/linux/ex04_lnx_ctrls.c
@@ -11,9 +11,9 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XSlider.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XSlider.h"
+#include "elem/XGauge.h"
 
 #include <time.h> // For clock() (frame rate reporting)
