+++ examples/linux/ex05_lnx_pages.c
@@ -11,8 +11,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
 
 #include <libgen.h>       // For path parsing
