+++ examples/linux/ex06_lnx_callback.c
@@ -11,8 +11,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
 
 #include <math.h>
 #include <libgen.h>       // For path parsing
