+++ examples/linux/ex09_lnx_radial.c
@@ -13,8 +13,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XGauge.h"
-#include "extra/XSlider.h"
+#include "elem/XGauge.h"
+#include "elem/XSlider.h"
 
 #include "unistd.h"   // For usleep() //xxx
