+++ examples/linux/ex10_lnx_textbox.c
@@ -10,8 +10,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XSlider.h"
-#include "extra/XTextbox.h"
+#include "elem/XSlider.h"
+#include "elem/XTextbox.h"
 
 #include "unistd.h"   // For usleep() //xxx
