+++ examples/linux/ex22_lnx_input_key.c
@@ -14,9 +14,9 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
-#include "extra/XSlider.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
+#include "elem/XSlider.h"
 
 #include <time.h> // For clock() (frame rate reporting)
