+++ examples/linux/ex24_lnx_tabs.c
@@ -12,8 +12,8 @@
 #include "GUIslice_drv.h"
 
 // Include any extended elements
-#include "extra/XCheckbox.h"
-#include "extra/XGauge.h"
+#include "elem/XCheckbox.h"
+#include "elem/XGauge.h"
 
 #include <libgen.h>       // For path parsing
