+++ src/elem/XTextbox.c
@@ -36,7 +36,7 @@
 #include "GUIslice.h"
 #include "GUIslice_drv.h"
 
-#include "extra/XTextbox.h"
+#include "elem/XTextbox.h"
 
 #include <stdio.h>
